train_tracker_project/requirements.txt
streamlit
pandas
python-telegram-bot==13.15
googletrans==4.0.0-rc1
pushbullet.py
